Add postgres only mode

+++ utils/datasets/destroy-dev-clone.sh
@@ -47,11 +47,11 @@ paused='false'
 # show function usage
 show_usage() {
     echo
-    echo "Usage: $0 -d {dataset_name} -n {instance_name} [-p {pause_tag}] [-t]"
+    echo "Usage: $0 -d {dataset_name} -n {instance_name} [-o {pg_dataset}:{pg_folder} ] [-p {pause_tag}] [-t]"
     echo
 }
 
-while getopts d:n:p:t opt; do
+while getopts d:n:o:p:t opt; do
     case $opt in
         d)  # Dataset name
             clone_dataset="$OPTARG"
@@ -61,6 +61,11 @@ while getopts d:n:p:t opt; do
             dev_name="$OPTARG"
             debug "instance_name: $dev_name"
             ;;
+        o)  # Only Postgres
+            pg_only="$OPTARG"
+            debug "Only cloning postgres at: $pg_only"
+            ;;
+
         t)  # Test mode
             test='true'
             debug "Running in test mode"
@@ -118,48 +123,50 @@ die () {
     exit $2
 }
 
+if [ "$pg_only" == '' ]; then
+    # Locate dataset info
+    clone_pool=
+    debug "Finding dataset source for $clone_dataset"
+    dataset_source=`dataset_source $clone_dataset`
+    o_source["${clone_dataset}"]="$dataset_source"
+    if [ "$dataset_source" == '' ]; then
+        die "Cannot locate source for $clone_dataset" 1
+    fi
+    debug "Found source at $dataset_source"
+    clone_pool=`echo $dataset_source | $CUT -d ':' -f 1`
+    
+    # Collect dataset folders
+    
+    rm -f ${TMP}/dataset_folders_$$
+    x=`$SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:folders ${clone_pool}/${clone_dataset}`
+    folders="$(echo -e "$x" | $TR -d '[:space:]')"
+    if [ "$folders" != '-' ]; then
+        NUM=1
+        while [ $NUM -le $folders ]; do
+            $SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:folder:${NUM} ${clone_pool}/${clone_dataset} 2>/dev/null  >>${TMP}/dataset_folders_$$
+            NUM=$(( NUM + 1 ))
+        done
+    fi
+    
+    rm -f ${TMP}/dataset_datasets_$$
+    x=`$SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:datasets ${clone_pool}/${clone_dataset}`
+    datasets="$(echo -e "$x" | $TR -d '[:space:]')"
+    if [ "$datasets" != '-' ]; then
+        NUM=1
+        while [ $NUM -le $datasets ]; do
+            $SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:dataset:${NUM} ${clone_pool}/${clone_dataset} 2>/dev/null >>${TMP}/dataset_datasets_$$
+            NUM=$(( NUM + 1 ))
+        done
+    fi
+    
+    ozmt_datasets=`cat ${TMP}/dataset_datasets_$$ 2>/dev/null`
+    
+    x=`$SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:postgres ${clone_pool}/${clone_dataset}`
+    postgres="$(echo -e "$x" | $TR -d '[:space:]')"
+    x=`$SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:postgresdev ${clone_pool}/${clone_dataset}`
+    postgres_dev="$(echo -e "$x" | $TR -d '[:space:]')"
 
-# Locate dataset info
-clone_pool=
-debug "Finding dataset source for $clone_dataset"
-dataset_source=`dataset_source $clone_dataset`
-o_source["${clone_dataset}"]="$dataset_source"
-if [ "$dataset_source" == '' ]; then
-    die "Cannot locate source for $clone_dataset" 1
-fi
-debug "Found source at $dataset_source"
-clone_pool=`echo $dataset_source | $CUT -d ':' -f 1`
-
-# Collect dataset folders
-
-rm -f ${TMP}/dataset_folders_$$
-x=`$SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:folders ${clone_pool}/${clone_dataset}`
-folders="$(echo -e "$x" | $TR -d '[:space:]')"
-if [ "$folders" != '-' ]; then
-    NUM=1
-    while [ $NUM -le $folders ]; do
-        $SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:folder:${NUM} ${clone_pool}/${clone_dataset} 2>/dev/null  >>${TMP}/dataset_folders_$$
-        NUM=$(( NUM + 1 ))
-    done
-fi
-
-rm -f ${TMP}/dataset_datasets_$$
-x=`$SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:datasets ${clone_pool}/${clone_dataset}`
-datasets="$(echo -e "$x" | $TR -d '[:space:]')"
-if [ "$datasets" != '-' ]; then
-    NUM=1
-    while [ $NUM -le $datasets ]; do
-        $SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:dataset:${NUM} ${clone_pool}/${clone_dataset} 2>/dev/null >>${TMP}/dataset_datasets_$$
-        NUM=$(( NUM + 1 ))
-    done
-fi
-
-ozmt_datasets=`cat ${TMP}/dataset_datasets_$$ 2>/dev/null`
-
-x=`$SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:postgres ${clone_pool}/${clone_dataset}`
-postgres="$(echo -e "$x" | $TR -d '[:space:]')"
-x=`$SSH $clone_pool zfs get -H -o value ${zfs_property_tag}:postgresdev ${clone_pool}/${clone_dataset}`
-postgres_dev="$(echo -e "$x" | $TR -d '[:space:]')"
+fi # pg_only
 
 ##
 # Locate and pause all related datasets
@@ -171,50 +178,52 @@ source clone-functions.sh
 # Destroy Clones
 ##
 
-if [ "$ozmt_datasets" != '' ]; then
-    for ozmt_dataset in $ozmt_datasets; do
-        this_source="${o_source[$ozmt_dataset]}"
-        o_pool=`echo $this_source | $CUT -d ':' -f 1`
-        o_folder=`echo $this_source | $CUT -d ':' -f 2`
-       
-        replication=`$SSH $o_pool zfs get -s local,received ${zfs_replication_property} ${o_pool}/${o_folder}`
-        if [ "$replication" == 'on' ]; then
-            timeout=900
-        else
-            timeout=60
-        fi
-        
-        result=1
-        SECONDS=0
-        while [ $result -ne 0 ]; do
-            $SSH $o_pool zfs list -o name ${o_pool}/${o_folder}/dev/${dev_name} 1>/dev/null 2>/dev/null
-            if [ $? -ne 0 ]; then
-                warning "Could not locate ${o_pool}/${o_folder}/dev/${dev_name}  Skipping destroy."
-                result=0
-                continue
-            fi
-            notice "Destroying ${o_pool}/${o_folder}/dev/${dev_name}"
-            if [ "$test" != 'true' ]; then
-                $SSH $o_pool zfs destroy -r -f ${o_pool}/${o_folder}/dev/${dev_name} 2>${TMP}/dataset_dev_destroy_$$.txt
-                result=$?
+if [ "$pg_only" == '' ]; then
+    if [ "$ozmt_datasets" != '' ]; then
+        for ozmt_dataset in $ozmt_datasets; do
+            this_source="${o_source[$ozmt_dataset]}"
+            o_pool=`echo $this_source | $CUT -d ':' -f 1`
+            o_folder=`echo $this_source | $CUT -d ':' -f 2`
+           
+            replication=`$SSH $o_pool zfs get -s local,received ${zfs_replication_property} ${o_pool}/${o_folder}`
+            if [ "$replication" == 'on' ]; then
+                timeout=900
             else
-                echo "TEST MODE.  Would run:"
-                echo "$SSH $o_pool zfs destroy -r -f ${o_pool}/${o_folder}/dev/${dev_name} 2>${TMP}/dataset_dev_destroy_$$.txt"
+                timeout=60
             fi
-            result=$?
-            if [ $result -ne 0 ]; then
-                warning "Failed to destroy ${o_pool}/${o_folder}/dev/${dev_name}"
-                if [ $SECONDS -gt $timeout ]; then
-                    error "Could not destroy ${o_pool}/${o_folder}/dev/${dev_name} aborting destroy-dev-clone." ${TMP}/dataset_dev_destroy_$$.txt
-                    die "" 1 
+            
+            result=1
+            SECONDS=0
+            while [ $result -ne 0 ]; do
+                $SSH $o_pool zfs list -o name ${o_pool}/${o_folder}/dev/${dev_name} 1>/dev/null 2>/dev/null
+                if [ $? -ne 0 ]; then
+                    warning "Could not locate ${o_pool}/${o_folder}/dev/${dev_name}  Skipping destroy."
+                    result=0
+                    continue
                 fi
-                sleep 60
-            fi
+                notice "Destroying ${o_pool}/${o_folder}/dev/${dev_name}"
+                if [ "$test" != 'true' ]; then
+                    $SSH $o_pool zfs destroy -r -f ${o_pool}/${o_folder}/dev/${dev_name} 2>${TMP}/dataset_dev_destroy_$$.txt
+                    result=$?
+                else
+                    echo "TEST MODE.  Would run:"
+                    echo "$SSH $o_pool zfs destroy -r -f ${o_pool}/${o_folder}/dev/${dev_name} 2>${TMP}/dataset_dev_destroy_$$.txt"
+                fi
+                result=$?
+                if [ $result -ne 0 ]; then
+                    warning "Failed to destroy ${o_pool}/${o_folder}/dev/${dev_name}"
+                    if [ $SECONDS -gt $timeout ]; then
+                        error "Could not destroy ${o_pool}/${o_folder}/dev/${dev_name} aborting destroy-dev-clone." ${TMP}/dataset_dev_destroy_$$.txt
+                        die "" 1 
+                    fi
+                    sleep 60
+                fi
+            done
+    
         done
-
-    done
-else
-    die "Could not locate any dataset listing for ${clone_dataset}  Make sure dataset's folder paramaters are set." 1
+    else
+        die "Could not locate any dataset listing for ${clone_dataset}  Make sure dataset's folder paramaters are set." 1
+    fi
 fi
 
 
@@ -225,13 +234,17 @@ fi
 if [ "$postgres" != '-' ]; then
     $SSH $p_pool zfs list -o name -H ${p_pool}/${p_folder}/${pdev_folder}/${dev_name} 1>/dev/null 2>/dev/null
     if [ $? -ne 0 ]; then
-        warning "Could not location ${p_pool}/${p_folder}/${pdev_folder}/${dev_name} Skipping destroy."
+        warning "Could not locate ${p_pool}/${p_folder}/${pdev_folder}/${dev_name} Skipping destroy."
         result=0
     else
-        notice "Destroying ${p_pool}/${p_folder}/${pdev_folder}/${dev_name}"
         if [ "$test" != 'true' ]; then
+            notice "Destroying ${p_pool}/${p_folder}/${pdev_folder}/${dev_name}"
             $SSH $p_pool zfs destroy -r -f ${p_pool}/${p_folder}/${pdev_folder}/${dev_name} 2>${TMP}/dataset_dev_destroy_$$.txt
             result=$?
+        else
+            echo "TEST MODE. Would run:"
+            echo "$SSH $p_pool zfs destroy -r -f ${p_pool}/${p_folder}/${pdev_folder}/${dev_name}"
+            result=0
         fi
     fi
     if [ $result -ne 0 ]; then
@@ -241,8 +254,8 @@ fi
 
 
 if [ "$pause" == "$$" ]; then
-    die "Destroying $clone_dataset complete" 0
+    die "Destroying $dev_name complete" 0
 else
-    notice "Destroying $clone_dataset complete"
+    notice "Destroying $dev_name complete"
     exit 0
 fi
